Added utilities for minimizing the size of the build-output

+++ init.sh
@@ -11,13 +11,32 @@ MAKEFILE=$(cat << MAKEFILE_END
 .PHONY: build test clean
 
 MAIN_FILE=$1.m4
+BUILD_FILE=build/$1.rkt
 
 build:
 	mkdir -p build
-	m4 \${MAIN_FILE} > build/$1.rkt
+	m4 \${MAIN_FILE} > \${BUILD_FILE}
+	
+	# Remove comments
+	sed -i '/;.*\$\$/d' \${BUILD_FILE}
+	sed -z -i 's/#|.*|#//g' \${BUILD_FILE}
+	
+	# Remove empty lines
+	sed -i '/^[[:space:]]*\$\$/d' \${BUILD_FILE}
+	
+	# Remove spaces around parens
+	sed -i 's/[[:space:]]*(/(/g' \${BUILD_FILE}
+	sed -i 's/)[[:space:]]*/)/g' \${BUILD_FILE}
+	
+	# Remove spaces in new-lines.
+	sed -i 's/^[[:space:]]*//g' \${BUILD_FILE}
+	
+	# The scary part -- replace newlines with spaces.
+	tr '\\n' ' ' < \${BUILD_FILE} > \${BUILD_FILE}.tr
+	mv \${BUILD_FILE}.tr \${BUILD_FILE}
 
 test: build
-	$(realpath ${TEST_RUNNER}) build/$1.rkt
+	$(realpath ${TEST_RUNNER}) \${BUILD_FILE}
 
 clean:
 	rm -r build
@@ -28,6 +47,7 @@ MAIN_FILE=$(cat << MAIN_FILE_END
 #lang plai
 
 include(../suplib/principles.rkt)
+include(../suplib/aliases.rkt)
 include(src/solution.rkt)
 include(../suplib/test-fw.rkt)
 include(test/tests.rkt)
@@ -54,8 +74,16 @@ echo "${MAIN_FILE}" > "$1/$1.m4"
 
 # Create the source directory
 mkdir -p "$1/src"
-echo "${SOLUTION_BOILER}" > "$1/src/solution.rkt"
+if test -f "$1/src/solution.rkt"; then
+    echo "Found an existing solution. Refusing to write boilerplate."
+else
+    echo "${SOLUTION_BOILER}" > "$1/src/solution.rkt"
+fi
 
 # Create the test directory
 mkdir -p "$1/test"
-echo "${TEST_BOILER}" > "$1/test/tests.rkt"
+if test -f "$1/test/tests.rkt"; then
+    echo "Found an existing test suite. Refusing to write boilerplate."
+else
+    echo "${TEST_BOILER}" > "$1/test/tests.rkt"
+fi
